+++ meerschaum/core/Pipe/_fetch.py
@@ -18,14 +18,14 @@ if TYPE_CHECKING:
     pd = mrsm.attempt_import('pandas')
 
 def fetch(
-        self,
-        begin: Union[datetime, str, None] = '',
-        end: Optional[datetime] = None,
-        check_existing: bool = True,
-        sync_chunks: bool = False,
-        debug: bool = False,
-        **kw: Any
-    ) -> Union['pd.DataFrame', Iterator['pd.DataFrame'], None]:
+    self,
+    begin: Union[datetime, int, str, None] = '',
+    end: Union[datetime, int, None] = None,
+    check_existing: bool = True,
+    sync_chunks: bool = False,
+    debug: bool = False,
+    **kw: Any
+) -> Union['pd.DataFrame', Iterator['pd.DataFrame'], None]:
     """
     Fetch a Pipe's latest data from its connector.
 
@@ -75,6 +75,8 @@ def fetch(
             if chunk_label:
                 chunk_message = '\n' + chunk_label + '\n' + chunk_message
             return chunk_success, chunk_message
+
+    begin, end = self.parse_date_bounds(begin, end)
 
     with mrsm.Venv(get_connector_plugin(self.connector)):
         _args, _kwargs = filter_arguments(
@@ -164,6 +166,6 @@ def _determine_begin(
         backtrack_interval = timedelta(minutes=backtrack_interval)
     try:
         return sync_time - backtrack_interval
-    except Exception as e:
+    except Exception:
         warn(f"Unable to substract backtrack interval {backtrack_interval} from {sync_time}.")
     return sync_time